requirements.txt
numpy==1.23.0
Pillow==9.2.0
scipy==1.8.0
scikit-learn==1.1.1
matplotlib==3.5.2
dlib==19.24.0
opencv-python==4.10.0.84
huggingface-hub==0.26.5
onnxruntime==1.17.0